:shirt: Move images into own namespace
