+++ src/Texim/Images/FullImage.cs
@@ -17,42 +17,23 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-namespace Texim.Pixels
+using Texim.Colors;
+
+namespace Texim.Images
 {
-    using System;
-    using System.Drawing;
-    using System.Drawing.Imaging;
-    using Yarhl.FileFormat;
-    using Yarhl.IO;
-
-    public class FullImage2BinaryBitmap :
-        IInitializer<ImageFormat>, IConverter<IFullImage, BinaryFormat>
+    public class FullImage : IFullImage
     {
-        private ImageFormat format = ImageFormat.Png;
-
-        public void Initialize(ImageFormat parameters)
+        public FullImage(int width, int height)
         {
-            if (parameters == null)
-                throw new ArgumentNullException(nameof(parameters));
-
-            format = parameters;
+            Width = width;
+            Height = height;
+            Pixels = new Rgb[width * height];
         }
 
-        public BinaryFormat Convert(IFullImage source)
-        {
-            if (source == null)
-                throw new ArgumentNullException(nameof(source));
+        public int Width { get; init; }
 
-            var image = new Bitmap(source.Width, source.Height);
-            for (int x = 0; x < source.Width; x++) {
-                for (int y = 0; y < source.Height; y++) {
-                    image.SetPixel(x, y, source.Pixels[y * source.Width + x].ToColor());
-                }
-            }
+        public int Height { get; init; }
 
-            var binary = new BinaryFormat();
-            image.Save(binary.Stream, format);
-            return binary;
-        }
+        public Rgb[] Pixels { get; init; }
     }
 }
