+++ src/Texim/Images/IFullImage.cs
@@ -17,16 +17,17 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-namespace Texim.Pixels
+namespace Texim.Images
 {
+    using Texim.Colors;
     using Yarhl.FileFormat;
 
-    public interface IIndexedImage : IFormat
+    public interface IFullImage : IFormat
     {
         int Width { get; }
 
         int Height { get; }
 
-        IndexedPixel[] Pixels { get; }
+        Rgb[] Pixels { get; }
     }
 }
