+++ src/Texim/Images/IndexedImageBitmapParameters.cs
@@ -17,17 +17,17 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-namespace Texim.Pixels
+namespace Texim.Images
 {
-    using Texim.Colors;
-    using Yarhl.FileFormat;
+    using System.Drawing.Imaging;
+    using Texim.Palettes;
 
-    public interface IFullImage : IFormat
+    public class IndexedImageBitmapParameters
     {
-        int Width { get; }
+        public ImageFormat Format { get; set; }
 
-        int Height { get; }
+        public IPalette Palette { get; set; }
 
-        Rgb[] Pixels { get; }
+        public IPaletteCollection Palettes { get; set; }
     }
 }
